+++ app/src/main/java/com/xpeho/xpeapp/enums/Screens.kt
@@ -7,4 +7,6 @@ enum class Screens {
     NewsletterDetail,
     Vacation,
     Colleague,
+    ExpenseReport,
+    ExpenseReportDetail,
 }