+++ app/src/main/java/com/xpeho/xpeapp/presentation/componants/AppBar.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -25,6 +26,7 @@ import com.xpeho.xpeapp.ui.theme.SfPro
 @Composable
 fun AppBar(
     title: String?,
+    fontStyle: FontStyle? = null,
     imageVector: ImageVector?,
     onTapBack: () -> Unit,
 ) {
@@ -38,6 +40,7 @@ fun AppBar(
                 fontSize = 24.sp,
                 fontWeight = FontWeight.W400,
                 fontFamily = SfPro,
+                fontStyle = fontStyle ?: FontStyle.Normal,
             )
         },
         navigationIcon = {
@@ -61,5 +64,6 @@ fun appBarPreview() {
         imageVector = null,
         onTapBack = {},
         title = null,
+        fontStyle = null,
     )
 }